refactor(aws): share neutral config loading
Centralize the ambient-provider isolation policy used by frozen credentials and the EKS client so copy-paste checks remain strict without diverging security behavior.
